Enable 2nd serial Port direct OctoPrint connection
This firmware change enabled the second serial port to directly connect to the RX/TX headers inside the case (usually used for the MKS WIFI module) to UART ports of a RaspberryPi running Octoprint. This should greatly increase the reliability of the connection. Aditionally some other features have been enabled that should increase performance of the printer.